+++ convert/main.c
@@ -119,9 +119,9 @@
 #include "common/box.h"
 #include "common/open-utils.h"
 #include "common/extent-tree-utils.h"
+#include "common/clear-cache.h"
 #include "cmds/commands.h"
 #include "check/repair.h"
-#include "check/clear-cache.h"
 #include "mkfs/common.h"
 #include "convert/common.h"
 #include "convert/source-fs.h"
